move accounting generator to tests
