+++ corehq/apps/accounting/tests/test_models.py
@@ -1,14 +1,13 @@
 import datetime
 from decimal import Decimal
-from django.db import models
 import mock
+
 from django.core import mail
-from corehq.apps.domain.models import Domain
+from django.db import models
 
 from dimagi.utils.dates import add_months_to_date
 
-from corehq.apps.accounting import generator, tasks
-from corehq.apps.accounting.generator import FakeStripeCard, FakeStripeCustomer
+from corehq.apps.accounting import tasks
 from corehq.apps.accounting.models import (
     BillingAccount,
     Currency,
@@ -19,11 +18,14 @@ from corehq.apps.accounting.models import (
     SMALL_INVOICE_THRESHOLD,
     StripePaymentMethod,
 )
+from corehq.apps.accounting.tests import generator
+from corehq.apps.accounting.tests.base_tests import BaseAccountingTest
+from corehq.apps.accounting.tests.generator import FakeStripeCard, FakeStripeCustomer
+from corehq.apps.domain.models import Domain
 from corehq.apps.smsbillables.models import (
     SmsGatewayFee, SmsGatewayFeeCriteria, SmsUsageFee, SmsUsageFeeCriteria,
     SmsBillable,
 )
-from corehq.apps.accounting.tests.base_tests import BaseAccountingTest
 from corehq.util.dates import get_previous_month_date_range
