+++ corehq/apps/accounting/tests/test_new_domain_subscription.py
@@ -1,11 +1,12 @@
 import datetime
-from corehq.apps.domain.models import Domain
-from corehq.apps.accounting import generator
+
 from corehq.apps.accounting.exceptions import NewSubscriptionError
 from corehq.apps.accounting.models import (
     Subscription, BillingAccount, DefaultProductPlan, SoftwarePlanEdition,
     SubscriptionAdjustmentMethod, SubscriptionType, EntryPoint,)
+from corehq.apps.accounting.tests import generator
 from corehq.apps.accounting.tests.base_tests import BaseAccountingTest
+from corehq.apps.domain.models import Domain
 
 
 class TestNewDomainSubscription(BaseAccountingTest):
