+++ corehq/apps/accounting/tests/test_renew_subscription.py
@@ -1,6 +1,5 @@
 import datetime
-from corehq.apps.domain.models import Domain
-from corehq.apps.accounting.tests.base_tests import BaseAccountingTest
+
 from corehq.apps.accounting import generator
 from corehq.apps.accounting.models import (
     Subscription,
@@ -8,6 +7,8 @@ from corehq.apps.accounting.models import (
     DefaultProductPlan,
     SoftwarePlanEdition,
 )
+from corehq.apps.accounting.tests.base_tests import BaseAccountingTest
+from corehq.apps.domain.models import Domain
 
 
 class TestRenewSubscriptions(BaseAccountingTest):
