+++ corehq/apps/accounting/tests/test_subscription_permissions_changes.py
@@ -1,18 +1,18 @@
-import os
 import json
+import os
+
+from django_prbac.models import Role, Grant
 
 from corehq import privileges
-from corehq.apps.accounting import generator
 from corehq.apps.accounting.models import BillingAccount, DefaultProductPlan, \
     SoftwarePlanEdition, Subscription, SoftwarePlan, SoftwarePlanVersion, \
     SubscriptionType, ProBonoStatus, SoftwarePlanVisibility, SoftwareProduct, \
     SoftwareProductType
-from corehq.apps.accounting.tests import BaseAccountingTest
+from corehq.apps.accounting.tests import BaseAccountingTest, generator
 from corehq.apps.app_manager.models import Application
 from corehq.apps.domain.models import Domain
 from corehq.apps.userreports.models import DataSourceConfiguration, \
     ReportConfiguration, ReportMeta, get_datasource_config
-from django_prbac.models import Role, Grant
 
 
 class TestSubscriptionPermissionsChanges(BaseAccountingTest):
