+++ corehq/apps/analytics/tests/test_subscription_properties.py
@@ -2,7 +2,6 @@ import datetime
 
 from django.test.testcases import TestCase
 
-from corehq.apps.accounting import generator
 from corehq.apps.accounting.models import (
     BillingAccount,
     DefaultProductPlan,
@@ -12,6 +11,7 @@ from corehq.apps.accounting.models import (
     ProBonoStatus,
     SubscriptionType
 )
+from corehq.apps.accounting.tests import generator
 from corehq.apps.analytics.signals import get_subscription_properties_by_user
 from corehq.apps.domain.models import Domain
 from corehq.apps.users.models import WebUser
